Keywordクラスにcreate_rankタスクの処理を移した
